+++ packages/multi-provider/src/provider.ts
@@ -138,7 +138,7 @@ export class MultiProvider<T extends Domain> {
    * @param nameOrDomain A domain name or number.
    * @returns A {@link Domain} (if the domain has been registered)
    */
-  getDomain(nameOrDomain: number | string): Domain | undefined {
+  getDomain(nameOrDomain: number | string): T | undefined {
     let name = nameOrDomain;
     if (typeof name !== 'string') {
       name = this.resolveDomainName(nameOrDomain);
@@ -154,7 +154,7 @@ export class MultiProvider<T extends Domain> {
    * @returns A {@link Domain}
    * @throws if the domain has not been registered
    */
-  mustGetDomain(nameOrDomain: number | string): Domain {
+  mustGetDomain(nameOrDomain: number | string): T {
     const domain = this.getDomain(nameOrDomain);
     if (!domain) throw new UnknownDomainError(this, nameOrDomain);
     return domain;