+++ packages/multi-provider/changelog.md
@@ -3,3 +3,5 @@
 ### Unreleased
 
 - refactor: additional functionality on the `Contracts` type
+- fix: multi-provider methods that return `Domains` now return the
+  `T extends Domain` type associated with the multi-provider